requirements.txt
requests>=2.31.0
supabase>=2.0.0
transformers>=4.45.0
torch>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
